core: Add mctp_pldm_tx()
Add a helper that inserts the correct "MCTP type" byte for a PLDM
message sent over MCTP. The message type byte is technically part of the
MCTP protocol (it's required to be after a MCTP message with the Start
flag set), but the current API requires all users to insert it manually.

This is tedious boilerplate at best so add a helper to the library to
handle it for us.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>